+++ ReleaseNotes.md
@@ -2,6 +2,8 @@
 
 - **[Breaking Change]** No longer reads standard input by default, adds a new `--read-stdin` flag to turn on this behavior (#752)
 - [Fix] Blog posts in the docs recipe no longer require Author or Category metadata (#751)
+- [Feature] New `Markdown.PrependLinkRoot()` method to prepend a configured `LinkRoot` to all `Markdown` links (#735, #748, thanks @tbolon)
+- [Feature] New `MarkdownPrependLinkRoot` setting for the blog and docs recipes to turn on this behavior for Markdown documents (#735)
 
 # 2.0.0